Working on ss
git-svn-id: svn://136.177.114.72/svn_GW/phreeqcpp/branches/ErrorHandling/src@6069 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ phreeqc/basicsubs.cpp
@@ -2794,7 +2794,7 @@ system_total_elt_secondary(const char *total_name)
 				/*
 				 *   Look for element
 				 */
-				for (int j1 = 0; j1 < (size_t) count_elts; j1++)
+				for (size_t j1 = 0; j1 < (size_t) count_elts; j1++)
 				{
 					if (strcmp(elt_list[j1].elt->name, total_name) == 0)
 					{
